Replace plugin-base with worker
Change-type: major
