+++ lib/contracts/loop-balena-io.ts
@@ -1,8 +1,9 @@
 import type { ContractDefinition } from '@balena/jellyfish-types/build/core';
 
-export const loopProductOs: ContractDefinition = {
-	slug: 'loop-product-os',
-	name: 'product-os',
+export const loopBalenaIo: ContractDefinition = {
+	slug: 'loop-balena-io',
+	version: '1.0.0',
+	name: 'balena-io',
 	type: 'loop@1.0.0',
 	data: {
 		roles: ['loop'],
