+++ lib/contracts/loop-company-os.ts
@@ -1,8 +1,9 @@
 import type { ContractDefinition } from '@balena/jellyfish-types/build/core';
 
-export const loopBalenalabs: ContractDefinition = {
-	slug: 'loop-balenalabs',
-	name: 'balenalabs',
+export const loopCompanyOs: ContractDefinition = {
+	slug: 'loop-company-os',
+	version: '1.0.0',
+	name: 'company-os',
 	type: 'loop@1.0.0',
 	data: {
 		roles: ['loop'],
