+++ lib/contracts/loop-product-os.ts
@@ -1,8 +1,9 @@
 import type { ContractDefinition } from '@balena/jellyfish-types/build/core';
 
-export const loopCompanyOs: ContractDefinition = {
-	slug: 'loop-company-os',
-	name: 'company-os',
+export const loopProductOs: ContractDefinition = {
+	slug: 'loop-product-os',
+	version: '1.0.0',
+	name: 'product-os',
 	type: 'loop@1.0.0',
 	data: {
 		roles: ['loop'],
