+++ lib/contracts/task.ts
@@ -2,6 +2,7 @@ import type { ContractDefinition } from '@balena/jellyfish-types/build/core';
 
 export const task: ContractDefinition = {
 	slug: 'task',
+	version: '1.0.0',
 	name: 'Task',
 	type: 'type@1.0.0',
 	data: {
