+++ lib/contracts/transformer-merge-properties.ts
