+++ lib/contracts/triggered-action-merge-draft-version.ts
@@ -1,9 +1,10 @@
 import type { ContractDefinition } from '@balena/jellyfish-types/build/core';
-import { TriggeredActionData } from '@balena/jellyfish-types/build/worker';
+import type { TriggeredActionData } from '@balena/jellyfish-worker';
 
 export const triggeredActionMergeDraftVersion: ContractDefinition<TriggeredActionData> =
 	{
 		slug: 'triggered-action-auto-merge',
+		version: '1.0.0',
 		type: 'triggered-action@1.0.0',
 		name: 'Triggered action for merging draft versions to final versions',
 		markers: [],
