+++ lib/index.ts
@@ -1,5 +1,5 @@
-import { JellyfishPluginBase } from '@balena/jellyfish-plugin-base';
-import { cards } from './cards';
+import { PluginDefinition } from '@balena/jellyfish-worker';
+import { contracts } from './contracts';
 import { actions } from './actions';
 
 export * from './types';
@@ -7,14 +7,15 @@ export * from './types';
 /**
  * The productOS Jellyfish plugin.
  */
-export class ProductOsPlugin extends JellyfishPluginBase {
-	constructor() {
-		super({
-			slug: 'jellyfish-plugin-product-os',
-			name: 'Product-OS Plugin',
-			version: '1.0.0',
-			cards,
-			actions,
-		});
-	}
-}
+export const productOsPlugin = (
+	definition: Partial<PluginDefinition> = {},
+): PluginDefinition => {
+	return {
+		slug: 'plugin-product-os',
+		name: 'Product-OS Plugin',
+		version: '1.0.0',
+		contracts,
+		actions,
+		...definition,
+	};
+};
