+++ package.json
@@ -46,7 +46,7 @@
   "dependencies": {
     "@balena/jellyfish-assert": "^1.2.11",
     "@balena/jellyfish-logger": "^4.0.27",
-    "@balena/jellyfish-plugin-base": "^2.2.25",
+    "@balena/jellyfish-worker": "^16.0.5",
     "lodash": "^4.17.21",
     "skhema": "^6.0.2"
   },
